+++ dnaStreaming/config.py
@@ -13,13 +13,15 @@ class Config(object):
     ENV_VAR_SUBSCRIPTION_ID = 'SUBSCRIPTION_ID'
     ENV_VAR_USER_KEY = 'USER_KEY'
     ENV_VAR_SERVICE_ACCOUNT_ID = 'SERVICE_ACCOUNT_ID'
+    ENV_VAR_OAUTH_TOKEN = 'OAUTH_TOKEN'
     ENV_VAR_API_HOST = 'API_HOST'
 
-    def __init__(self, service_account_id=None, user_key=None, config_file=None):
+    def __init__(self, service_account_id=None, user_key=None, oauth_token=None, config_file=None):
         self.customer_config_path = self.DEFAULT_CUST_CONFIG_PATH if config_file is None else config_file
         self.initialized = False
         self.service_account_id = service_account_id
         self.user_key = user_key
+        self.oauth_token = oauth_token
 
         self.headers = None
 
@@ -49,36 +51,70 @@ class Config(object):
     def get_authentication_headers(self):
 
         user_key = self.get_user_key()
-        if user_key:
+        oauth_token = self.get_oauth_token()
+        if oauth_token:
+            return {
+                'Authorization': f'Bearer {oauth_token}'
+            }
+        elif user_key:
+            # TODO: deprecate user key
             return {
                 'user-key': user_key
             }
-
         else:
-            msg = """
-                Unable to find credentials. Specify your account credentials by choosing one of the following ways:
-                - set your user key to an env var under the name 'USER_KEY'
-                - place your customer_config.json file in your home directory (e.g. ~/, $HOME/)
-                - pass the absolute path of your customer_config.json file to the constructor of the Listener class
-                - pass the user key as a parameter to the constructor of the Listener class
+            # TODO: deprecate user key
+            msg = f"""
+                Unable to find credentials. Please specify your account credentials using one of the following methods:
+                - Set your user key as an environment variable named '{self.ENV_VAR_USER_KEY}'.
+                - Set your OAuth token as an environment variable named '{self.ENV_VAR_OAUTH_TOKEN}'.
+                - Set the `user_key` property inside your customer_config.json config file (refer to README.rst for more details).
+                - Set the `oauth_token` property inside your customer_config.json config file (refer to README.rst for more details).
+                - Pass the `user_key` parameter to the dnaStreaming.listener.Listener class constructor.
+                - Pass the `oauth_token` parameter to the dnaStreaming.listener.Listener class constructor.
             """
             raise Exception(msg)
 
     def get_uri_context(self):
         host = os.getenv(self.ENV_VAR_API_HOST, self.DEFAULT_HOST)
         return host
+    
+    def get_oauth_token(self):
+
+        potential_oauth_token_var = lambda: self.oauth_token
+        potential_oauth_token_env = lambda: os.getenv(self.ENV_VAR_OAUTH_TOKEN)
+        potential_oauth_token_from_file = lambda: self._oauth_token_from_file()
+
+        potentials = [
+            potential_oauth_token_var,
+            potential_oauth_token_env,
+            potential_oauth_token_from_file
+        ]
+
+        # Lazily try to retrieve the value of the oauth token from all possible places where it may be set.
+        oauth_token = next((token for potential in potentials if (token := potential()) is not None), None)
+
+        return oauth_token
 
     # in the following two methods, note that we use "SERVICE_ACCOUNT_ID" as a legacy,
     # alternate name for the "USER_KEY" parameter, from the customer's perspective
     def get_user_key(self):
-        user_key = self.user_key if self.user_key else self.service_account_id
 
-        if user_key is None:
-            user_key = os.getenv(self.ENV_VAR_USER_KEY,
-                                 os.getenv(self.ENV_VAR_SERVICE_ACCOUNT_ID))
+        potential_user_key_var = lambda: self.user_key
+        potential_service_account_id_var = lambda: self.service_account_id
+        potential_user_key_env = lambda: os.getenv(self.ENV_VAR_USER_KEY)
+        potential_service_account_id_env = lambda: os.getenv(self.ENV_VAR_SERVICE_ACCOUNT_ID)
+        potential_user_key_from_file = lambda: self._user_key_id_from_file()
 
-            if user_key is None:
-                user_key = self._user_key_id_from_file()
+        potentials = [
+            potential_user_key_var,
+            potential_service_account_id_var,
+            potential_user_key_env,
+            potential_service_account_id_env,
+            potential_user_key_from_file
+        ]
+
+        # Lazily try to retrieve the value of the user key from all possible places where it may be set.
+        user_key = next((key for potential in potentials if (key := potential()) is not None), None)
 
         return user_key
 
@@ -107,3 +143,8 @@ class Config(object):
 
         return self.customer_config['subscription_id']
 
+    def _oauth_token_from_file(self):
+        if not self.initialized:
+            self._initialize()
+
+        return self.customer_config.get('oauth_token')